Closes: #981
